+++ src/assets/registerServiceWorker.js
