+++ src/index.js
@@ -1,21 +1,28 @@
-import React from 'react'
+import React, { Component } from 'react'
 import ReactDOM from 'react-dom'
+
 import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'react-router-redux'
-
-import Routes from './routes'
-
-// Static Files
-import './index.css'
-import registerServiceWorker from './registerServiceWorker'
+import { Route } from 'react-router'
 
 // Redux and Router
 import store, { history } from './modules'
 
+// Static Files
+import './index.css'
+import registerServiceWorker from './assets/registerServiceWorker'
+
+// Containers
+import { Top, ToDo } from 'containers'
+
 ReactDOM.render(
   <Provider store={store}>
     <ConnectedRouter history={history}>
-      <Routes />
+      <div>
+        {/* Routing */}
+        <Route exact path="/" component={Top} />
+        <Route path="/todo" component={ToDo} />
+      </div>
     </ConnectedRouter>
   </Provider>,
   document.getElementById('root')
